Implement Handlebars on UPDATE Restaurants
Added achievedRest event to allow update of restaurant on click of
glyph from handlebars.
Added check icon click function ui to allow single restaurant to be
updated when it's own trash icon is clicked.
Updated handlebars to allow removal of check icon restaurant on click of
glyph and replaced with Visited! message.
Next is formatting handlebars and adding events/ui for further function,
content, and style.